+++ pubrel.go
@@ -5,14 +5,16 @@ import (
 	"io"
 )
 
-// NewPubAck returns control packet with type PUBACK
-func NewPubAck() *PubAck {
-	return &PubAck{fixed: bits(PUBACK)}
+// NewPubRel returns control packet with type PUBREL
+func NewPubRel() *PubRel {
+	p := &PubRel{}
+	p.fixed = bits(PUBREL | 1<<1)
+	return p
 }
 
-// A PubAck packet is the response to a Publish packets, depending on
+// A PubRel packet is the response to a Publish packets, depending on
 // the fixed header it can be one of PUBACK, PUBREC, PUBREL or PUBCOMP
-type PubAck struct {
+type PubRel struct {
 	fixed bits
 
 	packetID   wuint16
@@ -21,7 +23,7 @@ type PubAck struct {
 	UserProperties
 }
 
-func (p *PubAck) String() string {
+func (p *PubRel) String() string {
 	return fmt.Sprintf("%s p%v %s%s %v bytes",
 		firstByte(p.fixed).String(),
 		p.packetID,
@@ -36,31 +38,31 @@ func (p *PubAck) String() string {
 	)
 }
 
-func (p *PubAck) AckType() byte {
+func (p *PubRel) AckType() byte {
 	return byte(p.fixed) & 0b1111_0000
 }
 
-func (p *PubAck) SetPacketID(v uint16) { p.packetID = wuint16(v) }
-func (p *PubAck) PacketID() uint16     { return uint16(p.packetID) }
+func (p *PubRel) SetPacketID(v uint16) { p.packetID = wuint16(v) }
+func (p *PubRel) PacketID() uint16     { return uint16(p.packetID) }
 
-func (p *PubAck) SetReasonCode(v ReasonCode) { p.reasonCode = wuint8(v) }
-func (p *PubAck) ReasonCode() ReasonCode     { return ReasonCode(p.reasonCode) }
+func (p *PubRel) SetReasonCode(v ReasonCode) { p.reasonCode = wuint8(v) }
+func (p *PubRel) ReasonCode() ReasonCode     { return ReasonCode(p.reasonCode) }
 
-func (p *PubAck) SetReason(v string) { p.reason = wstring(v) }
-func (p *PubAck) Reason() string     { return string(p.reason) }
+func (p *PubRel) SetReason(v string) { p.reason = wstring(v) }
+func (p *PubRel) Reason() string     { return string(p.reason) }
 
-func (p *PubAck) WriteTo(w io.Writer) (int64, error) {
+func (p *PubRel) WriteTo(w io.Writer) (int64, error) {
 	b := make([]byte, p.fill(_LEN, 0))
 	p.fill(b, 0)
 	n, err := w.Write(b)
 	return int64(n), err
 }
 
-func (p *PubAck) width() int {
+func (p *PubRel) width() int {
 	return p.fill(_LEN, 0)
 }
 
-func (p *PubAck) fill(b []byte, i int) int {
+func (p *PubRel) fill(b []byte, i int) int {
 	remainingLen := vbint(p.variableHeader(_LEN, 0))
 	i += p.fixed.fill(b, i)      // firstByte header
 	i += remainingLen.fill(b, i) // remaining length
@@ -68,7 +70,7 @@ func (p *PubAck) fill(b []byte, i int) int {
 	return i
 }
 
-func (p *PubAck) variableHeader(b []byte, i int) int {
+func (p *PubRel) variableHeader(b []byte, i int) int {
 	n := i
 	i += p.packetID.fill(b, i)
 	i += p.reasonCode.fillOpt(b, i)
@@ -81,14 +83,14 @@ func (p *PubAck) variableHeader(b []byte, i int) int {
 	return i - n
 }
 
-func (p *PubAck) properties(b []byte, i int) int {
+func (p *PubRel) properties(b []byte, i int) int {
 	n := i
 	i += p.reason.fillProp(b, i, ReasonString)
 	i += p.UserProperties.properties(b, i)
 	return i - n
 }
 
-func (p *PubAck) UnmarshalBinary(data []byte) error {
+func (p *PubRel) UnmarshalBinary(data []byte) error {
 	b := &buffer{data: data}
 	b.get(&p.packetID)
 	// no more data, see 3.4.2.1 PUBACK Reason Code
@@ -99,7 +101,7 @@ func (p *PubAck) UnmarshalBinary(data []byte) error {
 	return b.err
 }
 
-func (p *PubAck) propertyMap() map[Ident]wireType {
+func (p *PubRel) propertyMap() map[Ident]wireType {
 	return map[Ident]wireType{
 		ReasonString: &p.reason,
 	}