+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
+## [unreleased]
+
+- Remove tt.Interceptor
+
 ## [0.21.0] 2022-11-05
 
 - cmd/tt pub and sub commands work with mosquitto(anonymous)